Release version 1.2.2
This release is functionally identical in every way to version 1.2.1
(in fact the only things that's changed is metadata).

The idea here is to be able to close the legacy branch completely, which
had a few metadata changes. Additionally it also acts as a final
reminder that version 2 is imminent and may have breaking changes.

An additional upside here is that if we try to push this release out
after version 2.0.0, GitHub will show the releases in the wrong order,
even if we create the tag first, and my OCD will not allow that.