Include the ability to capture unknown overrides
This new API entry in the PackageInfo class captures all of the unknown
overrides, which are files that appear in the unpacked version of a
package, but not inside of the `sublime-package` file.

Such files are not overrides because they're not overriding anything,
but we would still like to fetch them for use in other places, such as
the OverrideReport, where we can report that they're not overrides but
still doing something a user might not expect.

This is an API function even though the information is available through
existing API methods (by getting the set of files and the set of 
overrides, which is all this does), but this allows us to cache the
value as well.