Update the syntax to highlight unknown overrides
For now, this is done by including a `[?]` prefix onto unknown overrides
in a manner similar to how `[X]` is used for expired overrides.

For the moment, this mark is styled as a comment to make it more 
distinct from overrides. This may change; we perhaps want the whole of
the file to be marked this way so that it appears even more distinct,
although this may interfere with the ability to edit such files through
the context menu.

Along with the syntax changing, the syntax test is also updated as well.