Make the override report record unknown overrides
In order to be able to exclude unknown overrides from some commands, we
need to know what they are. However this information is only available
when there is a PackageInfo() for the package, which is an expensive get
if you don't have one.

Since the Override report already knows about these files when it
generates the report, we can store that data in settings for later
examination if desired.

We do that by storing a new setting as a dictionary keyed by package
names that reference lists of files that are unknown.