Include a new context help for detecting unknowns
This adds to the ContextHelper class the ability to detect when an
override is unknown, by passing in a view that contains an override
report and a context that provides the information on the package and
override to check.

This consults the newly added setting and provides a boolean result that
indicates the status of that override.