Simplify the report generation
As a simplification on report generation, we gather the complete list of
unpacked files and iterate over them when we're generating the report.
Since we would otherwise be basically reconstructing this list by adding
the overrides back to the list of unknowns.

This also points out that in this particular instance, the new API for
getting this information is not needed since the report could synthesize
it. However, I'm leaving it in because it adds a nice bit of
functionality that you would otherwise have to write yourself, and we
might need it again later.