Extend the bulk diff report for unknown overrides
This extends the calculation of diffs in the bulk diff report so that
it will show unknown overrides in the report, but report that it can't
diff them because there is no base file.

This enhances the bulk diff report in a manner similar to what was done
for the override report and makes it easier to vet package changes when
those changes include files that don't exist in the base package, such
as when extending a package.