Extend the package file contents method
Previously, the method in the PackageInfo class for getting the contents
of an override base file would return the content back as a list of
lines in the file, since that is how the diff functionality in difflib
wants it.

If we're going to revert a file back to the original, that should
include the original line endings as well. For this reason, we now allow
this method to grab the actual raw content of the file instead of
splitting it into lines.

This is now also used in the external diff functionality to ensure that
the file that we create is identical to the internal file, as it was
using this method previously and writing files a line at a time in a 
way that would indeed cause the output to potentially have different
line endings.