Include a new command for reverting overrides
This implements a new context menu command for reverting an override
back to it's pristine state. Outwardly this is similar to removing the
override, but keeps the file in place for future edits.

This is implemented as a context menu command on overrides that are not
unknown (i.e. they have backing files).