Fix display issues for packages with only unknowns
There was an issue in the code for the new handling of unknown overrides
that would cause display issues if unknown overrides were the only files
in the unpacked package.

In the case of the override report, the short circuit for there being no
overrides would eject out of displaying anything, which is now fixed.

In addition, the code for the diff report would still work, but it would
output a line about no simple overrides found, which was a bit
unpleasing to the eye.

This fixes both problems by having those messages only appear when there
are no simple OR unknown overrides, which better mimics the previous
display. When there are only unknowns, we no longer say that there are
no simple overrides (which is technically still true).