Include a new command to create overrides
As the first instance of new functionality using our new browse code
this adds a new command that allows for the creation of a new override.

This prompts you for a package and then a resource from within that
package that isn't already an override and opens a temporary buffer for
that resource.

Saving the buffer will create an override. This works similar to how the
`PackageResourceViewer: Open Resource` command works except that the
resulting file properly propagates the content of the file for mini_diff
purposes, which makes tracking what you did to the file easier while
working on it.