Make the new creation command more robust
This adds some simple tweaks to the new creation command to make it a
little more robust. This is evenly split by making sure that we don't
try to update the settings in the new view until the buffer fully loads,
but also the command name once a resource picked was not properly
updated during the code move.