Allow for custom package filters while browsing
This enhances the PackageBrowser class (and the PackageResourceBrowser
class that uses it) so that it can take an optional filter function that
is used to select the packages to display instead of using the file type
critera provided.

This could also be implemented by forcing the filter to happen at the
point where the package list is provided, but that makes things a bit
more complicated to use since it's likely that code will want to
leverage the PackageResourceBrowser to create the package list directly.

This new functionality is used in the new CreateOverride command to
filter away packages that are only unpacked, since they can't possibly
be used to construct an override.