Enhance the code for creating an override
This is a slight tweak to the code to make it a little bit cleaner.
Instead of creating first a version of the full resource using Packages/
for testing if the resource exists, then again with ${packages} to open
it, we create a shorter stub version first and use some in-place
concatenation to save some redundant work.