Add a status method to the PackageInfo class
This method returns a dictionary that contains the core statistics of
the package and is meant for cases where package information needs to be
persisted for later use without storing the actual PackageInfo class
instance itself.

Our use case here is in being able to store package information in
report settings for powering package related functionality, since the
API does not allow for storing arbitrary class information in settings.