Persist package information in report views
Using the new status API added for to the PackageInfo class we now store
the package information for all packages that appear in all of our
reports.

The Package List report uses the non-detailed API since it doesn't do a
full override scan when it generates, but the other reports store more
detailed status since it's essentially available "for free" due to the
information already being collected as a part of the report generation
process.