Include Package Control metadata in status fields
Any package or dependency that Package Control installs has a metadata
file injected into it's package that tells Package Control the
information about that package that it needs to manage the package.

The PackageInfo class has been augmented so that the new status() method
contains an entry named "metadata" which is the contents of this
metadata file. 

As a part of loading the metadata, the dependency information (for a 
package) is gathered and injected into the metadata. When the actual
package metadata does not contain any dependency information but there
is a dependencies.json file, that file is loaded and injected into the
metadata directly as if it was always there.

When present, this provides information like a package description and
version, amongst other information; it will be an empty dictionary for
any package not installed by Package Control.