Implement a PackageInfo file API
This introduces a new API to the PackageInfo class that allows you to
determine if a particular resource appears in a package or not (based on
it being in the package zip file or on disk) as well as being able to
load the contents, either as a string or as a binary blob of bytes.

This is an implementation of the Sublime API for finding resources and
loading them that doesn't rely on the Sublime file catalog, which allows
us to load package resources even for packages that are ignored.