Extended metadata loading in PackageInfo
This change brings an enhancement to the existing mechanism for loading
in the PackageControl metadata files for packages by loading in the
extra dependencies.json file that many packages use instead of having
their dependencies declared in the channel file.

The upshot of this is that the metadata information can now contain
useful dependency information for more packages, which will enhance the
display of dependencies in our upcoming hover popups.

In order to support packages that are manually installed but still have
dependencies information, the loading of the dependency information will
happen even for a file that does not contain any metadata, UNLESS the
metadata is broken (i.e. non-parseable JSON).