Implement initial package hover popup support
This includes the initial implementation of a package hover popup for
reports that display package names. The popup will display information
on the package, including it's dependencies and version by using the
metadata gleaned from the package file.

This still needs some work; the styles need to be changed and tweaked,
the wording needs to be cleaned up in some places. We also want to
augment the PackageInfo class so that Shipped packages have metadata
information as well.