Detect dependencies that are in installed packages
Our original test for whether or not something is a dependency is based
on checking for an unpacked metadata file, but that is technically not
correct.

In the general case, Package Control is what installs dependencies and
it installs them (currently) as unpacked packages in the Packages folder
and places a dependency-metadata.json file to provide metadata
information.

As it transpires, Package Control considers it's
0_package_control_loader  bootstrapping package as a dependency (i.e. it
contains the metadata file) but it's stored as a sublime-package file.

Here we shift the detection for whether or not a package is a dependency
until after all of the package info is loaded and then check for the
presence of the dependency information, packed or not. This allows us to
detect the package control loader as a dependency, as we should.