Include some custom metadata for shipped packages
This includes some default metadata information for packages that ship
directly with Sublime so that the package popup can provide more useful
information for those packages. Along with that is a new internal method
for being able to grab this meta information and adjust it as needed.

The method can return overall default information for packages and
dependencies, as well as being able to infer version information for
shipped packages (based on the Sublime Text version), URL's for all of
the default packages that appear on GitHub as well as some default
descriptions.