Include some core filtering for unknown overrides
This implements some common core code usable throughout the code base
for filtering away some or all of the unknown overrides.

This works by having a helper function to obtain the value of the
setting and return back a list of precompiled regular expression filters
from the settings along with another helper to do the match.

The filtering happens by being given a list of all unpacked files and a
list of which of those are known to be unknown overrides, and then
removing from the package file list all of the entries for unknown
overrides that match one of the patterns.

This allows display code to iterate over the whole package contents and
skip over any unknowns that the user doesn't want to see.