Reorganize code for ignoring unknowns
In order to be able to display not only the number of unknown overrides
in hover popups but also the count of them that have been ignored by
the appropriate setting, this refactors the recently added code to shift
the handling for doing the filtering into the PackageInfo class.

This allows that class to keep track of the last count of filtered files
so that it can be reported via status() and thus used by external code
as well.