Factor out new override view setup code to core
The code from override_audit_create_override that is responsible for
setting up a new view to be an override has been factored out from that
command and into the core so that it can be re-used in an upcoming
command.

As a part of this, the method gains the code that sets the temporary
setting that marks a view as being a new override so that the event
listener can track it, to save having to manually do that later as well.