Add context command to create overrides from views
This adds a new command that can be invoked in existing views to promote
that view to the status of a new override similar to how the existing
command does it.

For the command to activate, the view it executes in must have a file
name set that is based in the Packages folder and which does not exist
on disk but is a file that exists in the underlying package file. The 
file must also be read-only as well (e.g. opened from 'View Package
File' in the command palette).

When invoked the view is converted to be read/write and the temporary
setting is applied which makes the buffer respond to save events by
creating an override, as the existing command does.