Hide ignored packages from Create Override command
As a bit of an oversight, the Create Override command would offer
packages which are currently disabled to create overrides inside of.
This is technically possible to do, but the current implementation will
fail because it relies on Sublime to open the package file. 

This can be revisited in the future if needed (perhaps it is handy to 
disable a package before making modifications), but for now we will
leave things such that it's only possible to create overrides in enabled
packages.