Include a new mini_diff setting for overrides
Adding in a new setting `mini_diff_underlying` which is used when
opening or creating an override to set up the `mini_diff` functionality
so that the base document is the underlying content of the override.

When set to `true`, every time an override is loaded or saved, or when
the Create Override commands are used, the backing document of the view
is reset to the packed content of the file being edited so that the
mini_diff in the gutter can be used to show you exactly what is changed.

When set to `false` or when the `mini_diff` setting in the user
preferences is not set to `true`, the `mini_diff` in files is not
touched and whatever would normally happen with the file happens
instead.

For new overrides that means no `mini_diff` (Sublime disables it for the
view because there is no file on disk) and for overrides that are a part
of a git repository, the git status would be used if the `mini_diff`
setting is `true` or `"auto"`.