Enhance check_potential_override to get content
This enhances the check_potential_override method in the PackageInfo
class so that it has a new argument that gets it to return back the
content of the underlying file for anything that it thinks is an
override.

The new argument defaults to what the function did before for
compatibility, but when enabled (and the other argument is also True),
along with the package and override name, the tuple will also contain
the content of the file as it exists in the `sublime-package` file.