Implement new setting by enhancing event listeners
This ties all of the prior commits together by using the new setting and
new changes to the check_potential_override() method to set the
reference document of the view as appropriate.

With this in place, when the setting is set the mini_diff will track the
underlying file, which lets you use the internal Sublime functionality
for navigating diff hunks and reverting them.