Update the README and CHANGELOG for new features
As always I remain incapable of remembering that I did not update the
CHANGELOG or README file for new features until after they have been
pushed up for others to see, and am forced to show my shame with a loose
commit in `master` that covers it instead.

This tweaks the existing CHANGELOG entry for creating an override to
point out the new version that is available in the context menu for
read-only resources opened by `View Package File` and outlines the new
setting that controls the `mini_diff` functionality.