Update the menus and command palette
After some thought, the newly added commands are now also added to the
main menu and the command palette. In addition the ellipsis in the menu
file has been swapped for the appropriate UTF-8 character instead of
three consecutive periods.

The README has been adjusted for the new commands as well.