Remove the settings proxy
The proxy was used to make OverrideAudit transparently display either a
single or multiple entries for settings and key bindings in the menu
depending on the version of Sublime in use, since older versions do not
have an `edit_settings` command and worked differently.

Now that we only support newer versions of Sublime in version 2, this is
no longer needed for compatibility reasons.