Small cleanup in the package popup generation code
This cleans up the code for generating the package popup to make it a
little bit clearer, adjust some of the text, and generally make things
somewhat better.

Overall this is probably still a bit hack and perhaps there is a better
way to generate HTML conditionally without being overly heavy weight for
something this simple, but it will do until popups on overrides are
added.