Enhance context override creation to include pkg
The `override_audit_context_create_override` has been extended so that
it will enable itself when it's invoked on the name of a package in a
report and that package can potentially create overrides.

When invoked in this manner, the package name is pre-populated so that
the user only has to enter the override that they want to create.

To make this and another upcoming bug fix easier, the context helper now
has a method for determining if it's being invoked on a package that can
conceivably contain overrides.