Fix bulk diffing to only apply for valid packages
During the migration of commands from 1.x to 2.x, the command for
executing a bulk diff of a package as a context item got simultaneously
smarter and stupider.

In particular, it used to be possible to try and bulk diff any package,
even if it didn't have any overrides in it. During the upgrade, some
work was done to try and make the command only apply when there was an
unpacked folder for the package.

However this is a bit dodgy because it also triggers for unpacked only
packages (like `User`) and the implementation didn't agree with what
is_visible said, so the command would appear disabled instead of being
hidden.

This is now changed so that both agree and it only enables itself when
the package has both a package file AND unpacked files, which seems like
a nicer implementation overall. This wasn't previously possible because
there was no cached package information to rely on, but now there is.