Add a temporary setting to disable hover popup
In the upcoming release of OverrideAudit the hover popup is still
undergoing some revisions as to what it should contain and how it should
look.

Until things stabilize, the `enable_hover_popup` setting can be used in
the OverrideAudit.sublime-settings file to disable the hover popup
entirely.

This setting may go away in the future once things are stable, so it is
not being documented currently under the theory that not documented
means impermanent.  In the future if we decide to keep it, it will be
added to the documentation.