Add new link to the hover popup
This tweaks the working slightly on some text in the popup and also adds
a new potential link for creating an override in a new package directly
from the hover popup.

The hover popup is very much still a work in progress, so this may need
to be cleaned up or possibly removed completely or otherwise visually
changed.

In the meantime, the links have been moved to their own section; each
link only shows itself when it is valid, and the entire section is
hidden if it does not apply.