Improve the confirmation message for reversion
Unlike deleting an override (which sends the file to the trash) or
freshening one (which only modifies the file stamp), reverting an
override will cause permanent data loss for anyone that hasn't kept a
copy of the modified resource laying around.

For that reason, the content of the dialog asking you to confirm the
revert operation spells this out specifically now as an extra reminder
that someone bad could be about to occur.