Syntax definition addition and cleanup
In preparation for some future work, this implements some cleanup of our
syntax definitions and includes a new one. In particular, the names of
existing syntaxes have been modified to better refer to what they are
instead of having the generic name "OverrideAudit".

This also includes a general file rename for the Diff Report syntax,
which is the only one of the existing report syntaxes to not follow the
proper naming convention.

Also added are a new syntax for our diffs and patches. This is currently
done via an embed on the built in diff syntax, with a different top
level scope and first line/extension matches. These will be used for
some upcoming changes regarding patches.