Set up for new patch creation logic
This adds a new stub command that will be responsible for triggering the
creation of a new patch, as well as a context sensitive version that is
enabled in reports, overrides and override diffs. This is also available
in the main menu and in the command palette.

Still to do are having the command prompt you for missing arguments in
an interactive fashion and actually create the patch, but this sets up
for being able to more easily trigger things for testing. It also helps
to get a handle on whether we're on the right track for where and how
the command should be visible.