Include a new setting for the patch location
In our current design for the patch feature, it will be possible for a
patch to contain a binary file (e.g. a theme). In such a case we need to
create  a sidecar file (simple zip) along the textual patch file.

We could generate the textual patch into an unsaved file tab for the
user to save where they want, but that would require keeping track of
the sidecar in a temporary location and also moving it if/when the patch
gets saved.

For this initial version, we will instead have a default path configured
for saving the patch (similar to how Package Control generates packages)
and we can then use that to prompt the user by providing them the
ability to modify the path as needed.

This change includes the setting as well as the support code for
determining the base filename and path for patch output. In addition it
introduces a new file where all of the patch related code will be
stored, since that is a more logical location than the core and allows
us to  group all of the similar code together.