Modify the PackageBrowser to return pkg_info
In the original implementation of the PackageBrowser class, the callback
would get invoked with either the name of the package selected by the
user or None.

This changes the callback so that it instead gets invoked with the
PackageInfo instance of the selected package, or None.

This makes it easier on code that wants to do something with a selected
package because it ends up with the package info it probably wanted to
use anyway (and can still get the name from it if that's all that is
required).