Include an initial version of a binary diff
As a WIP, this includes a small portion of Lib/base64.py from the Python
library source code to bring the base85 encoding/decoding added in
version 3.4. This is used to inline an encoded version of a binary file
that appears in an unpacked file, so that it can later be recreated.

The ultimate idea is that this will be configurable and we would either
inject these files like this, or use a smaller chunk (as in the initial
stub code for this) and attach the file to a sidecar file that would be
generated along with the patch.

We're not doing any diff at the moment; any binary file that exists
unpacked but not in the source package, or for which the CRC is
different is inlined in its entirety.