Include a release checklist file
This includes a simple release checklist file for the various files that
need to be updated (or verified) during a new release of the package as
an extra verification that we're not forgetting anything important.

This file is excluded from the git archive so that when the package is
installed by the end user they don't end up with a version of the file,
since it's likely not interesting to the end user how we manage our
releases.