Add the ability to annotate overrides in browser
This enhances the ResourceBrowser and PackageResourceBrowser packages to
annotate overrides in the display with text that marks them as overrides
so that when choosing resources, you can differentiate if needed.

This is controlled by a flag which defaults to False for compatibility
with previous calls, and only takes effect when you specify the resource
type of ALL.

This is because for OVERRIDES everything is already an override, and for
NONOVERRIDES you'd only see files that are not overrides anyway.