Allow create override to allow existing overrides
The command for creating overrides was already smart enough that if
given an explicit package and file and that file already exists on disk
that the file should just be opened instead of taking the steps of
preparing for it to be an override.

Given that, the command was given an extra argument that tells it if it
should include existing overrides in the display (the default is not to
do that), and in doing so allow you to open any resource, potentially
also creating an override out of items as PackageResourceViewer would
do.

To do this, the arguments to the PackageResourceBrowser have been
modified as appropriate.