Add python runtime version to the package info
The PackageInfo class now carries a string that indicates what version 
of the plugin host the plugins in that package run in. This is initially
set to the empty string as an indication of no plugins.

When loading metadata, we check to see if there are any plugins defined
in that package, and if there are we populate the member with the
appropriate version of python, defaulting to 3.3 but loading the content
of the .python-version file if there is one.

This information is injected into the status information that reports 
use to stash information.