Display python runtime in the package popup
This displays the Python version of the package in the package hover
popup so that it's immediately obvious in what runtime that package will
run. This information is taken from the new status information which has
been added.

If the python version is the empty string (which is the default) then
no python version is displayed. This string should only be the empty
string for packages that don't have plugins in them, since as soon as we
detect that there's a plugin we change the default to the legacy 3.3
version.