Use custom context menus in reports
In report views, we now apply our own custom context menu that only
shows relevant commands, which are OverrideAudit commands only.

In order to make this more seamless, commands have been modified so that
their description methods know if they should include
the "OverrideAudit: " prefix or not; they will not when used in a report
but they will in other cases.