Fix false positives in the Python plugin detection
The code for detecting whether or not a Package contains any plugins was
falling afoul of the fact that syntax test definitions take the
extension of the type of file that they're testing.

In particular, the Python package contains a syntax test which was
erroneously detected as being a syntax test. That package currently
contains no plugins, so not only is this a false positive, but it also
reports the "wrong" Python version (since all shipped packages with
plugins generally run in the new host).

This is fixed by specifically excluding any plugin file in the root of
the Python package that starts with `syntax_test_`,since that is
required of the syntax tester (per the documentation).

This required a mild refactor of the existing detection so that it could
be easily applied to both detection mechanisms.