Implement a stub for freshening all overrides
This includes the stub core code for a new command that will attempt to
freshen all expired overrides that are visible in the current report.

Currently the command does nothing, but it will appear in the main menu,
context menus and in the command palette as long as the current tab is a
view that has at least one expired override in it.