Freshen thread update to allow names or objects
Previously the thread for freshening expired overrides assumed that you
would give it the name of a package and it would create a PackageInfo()
instance out of it.

This simple modification allows the method to also be given a
PackageInfo() instance directly. This allows external code that has
already gone through the work of determining the content of a package to
re-use that information instead of forcing another collection.