Allow for multiple freshens at one time
This (partially complete) change enhances the freshen thread and its
external interface so that it supports the notion of freshening the
overrides in multiple packages at once, or just multiple overrides at
once.

This is functionally working, but prompting and status messages need
more work in order to be release ready.