Allow freshen thread to work viewless
To allow for freshening of overrides without there being a ready-made
report view present, this modifies the freshening thread so that if you
don't pass in a view, it doesn't complain and instead just doesn't try
to update the view.

This allows it to be used by code that may want to freshen overrides
without generating a report first.