Implement the initial version of Freshen All
This provides the simple implementation of freshening all overrides that
currently exist in the given report by using the newly enhanced freshen
thread to run the freshen over multiple packages at once.

This still needs work since the interface to the freshen thread is still
being worked on. For example, it still needs the ability to constrain
the freshen to unmodified files (as determined by a diff), which should
be reflected here but currently is not.