Release version 2.2.1
This is a small bugfix release to patch up a defect in the newly added
`Open Resource` command. The intent of the newly added
`include_existing` argument was to allow for any package resource to be
added, but the package filer was still incorrectly filtering away
packages that were not packed (since only those packages can contain
overrides).

This fixes that by applying a different filter that only ensures that
disabled packages are filtered away.

Thanks to @keith-hall for the bug report.