Fix issues with AutomaticPackageReloader
This factors away some code that used to appear as a class method in the
PackageInfo class into a helper function instead. The method in question
was used to set up a variable used to store the location for shipped
packages.

Due to the way that APR reloads packages and the way that the code was
structured, hot-reloading the package would cause the shipped package
path to be lost.

The code has been moved out into a helper method instead, which lazily
captures the value on first call.