Update cursor location on report create/refresh
As previously written, when creating or refreshing any report the cursor
would be left at the last line of the report, due to the `append` that
is used to populate the report.

While this works, it's mildly annoying that on load if you try to move
the cursor the view immediately shifts to the bottom of the view, where
the cursor is located.

This change puts the cursor at 1:1 on create/reload, which more closely
mimics how opening a file would act.