Add ability to hide unchanged from override report
As a first addition relating to #35, this adds a new argument to the
command that generates override reports that causes it to not display
any overrides that diff out to be identical to the underlying file.

This is currently exposed as a new entry in the command palette and the
`Tools > OverrideAudit` menu to generate a report with the new option,
and it also persists in the report so that a refresh will keep the same
option.

Currently this option does not stop the display of binary files that may
or may not be similar (they are currently not diffed or otherwise 
compared to the underlying file) and it also leaves expired overrides in
the list regardless of their content.