Add ability to hide unchanged from package diffs
Similar to the previous commit, this extends the commands responsible
for generating bulk diff reports (single package or multi-package) to
include an extra argument that tells the command to elide unchanged
files from the report.

Any given file will not be included in the diff report if it appears to
be unchanged, though unknown and binary files will appear. 

In the event that the contents of a package contain only unmodified
files, that entire package will be dropped from the report entirely.

For this reason, the Bulk diff reports now generate a warning at the top
to indicate to you in this situation that other files may exist that are
unchanged, to remind you that you're not seeing everything there is to
see.