Fix diffs of packages that do not support diffing
This fixes some (mostly hidden) defects in the bulk diffing code that
would trigger an error in cases where the package is entirely
un-diffable:

  * When there is only a `sublime-package` file but no unpacked folder
  * When there is an unpacked folder but no `sublime-package` file

Both could cause the diff code to throw an exception instead of
gracefully handling the issue.


This is mostly hidden because the outer command is normally used in a 
way where it prompts you for the package(s) to diff, and in such a case
it would not allow you to choose a package that would cause this problem
to surface.

As such, the bug fix is to allow for others to invoke the command and
get a meaningful result rather than an exception without their having to
know ahead of time if the package is safe or not (which is a non-obvious
check for outside use cases).