Fix an issue in filter_unmodified_overrides
As implemented, the method was taking a set of overrides and modifying
it in place to come up with a filtered list, but the intention 
(and call) was to assume it returned a new list.

This resolves that by making a duplicate set for modification and return
so that the incoming list is left untouched. This resolves what would be
a bug in an upcoming change, since we're going to call this with the
list of expired overrides that's cached in a PackageInfo instance.