Flesh out our freshen command more
This is closer to a final implementation of the freshen all command. It
freshens either literally every expired override, or only those that are
currently unmodified but still expired, with the default being the
latter for safety. 

The current plan is that the addition of a command palette entry to
invoke the command for a full unfettered refresh will be an exercise for
the more advanced user to undertake on their own so that the default
experience for newer users is safer. This may instead be gated behind
an "advanced_mode" type setting.

In use this uses the code from the existing commits but performs actual
checks to determine what should be freshened.

This brings the code closer to an implementation of #34, with the
introduction of a command that can take the action manually in a report,
though more work is still required before this can be considered final.