Fix view index errors if images are open
Our entries in the tab context menu could cause errors to be thrown
in the console (and the command to be unavailable in the menu) if the
tab prior to a report or an override was a non-`TextSheet`, such as an
image.

This is because our command gets told what sheet index it is in the
group, but we were using `window.views_in_group()` to look up the group.

We now use `window.sheets_in_group()` instead, so that the index we're
given is always valid.