Add new custom context menu for freshen all
This uses the newly added (on the branch we just rebased this one onto)
`caption()` method in our description method so that if the command is
used in the custom context menu, it will not include the package prefix
on the name.

We also add the command to the custom report menu as well.