Fix context menu items in the tab context menu
In ST3 the only time the `event` argument gets provided to a command is
when it's invoked from the context menu inside of a view (and it
actually requested to get the event).

In ST4, the `event` argument will also be raised on commands in the tab
context menu to provide keyboard modifier key information.

The context commands were running afoul of this in Sublime Text 4
because they ask for an event but expect it to always contain position
information.

This simple fix only assumes that an event should be handled if it
exists and also contains positional information.