Fix Open Resource command palette entry not showing all files
In the command palette, OverrideAudit: Open Resource executes command: override_audit_create_override {"include_existing": true}
Prior to this commit, it was always firing the PackageResourceBrowser with "unknown" set to `False`.
But, as the command palette entry's name indicates, it would be expected to be able to select *any existing* resource from this list - whether it only exists as an extra file in the package or not. (This is important when creating new syntax definitions which extend other ones, for example - they won't exist in the original package, but we still want to see them and open them.)
The solution therefore is to set `unknown` to the value of `include_existing`, to keep it working the same way for other use cases,
while fixing the aforementioned command palette entry.